pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexuseval"
version = "0.4.0"
description = "A production-ready LLM evaluation framework with intelligent caching and dataset management."
authors = [{name = "Shubham Salokhe", email = "shubhamsalokhe@ymail.com"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["llm", "evaluation", "hallucination", "rag", "ai", "testing", "openai"]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

dependencies = [
  "openai>=1.0.0",
  "tqdm>=4.0.0",
  "numpy>=1.20.0",
  "pydantic>=2.0.0"
]

[project.optional-dependencies]
providers = [
  "anthropic>=0.39.0",  # For Claude models
  "google-generativeai>=0.8.0",  # For Gemini models
  "groq>=0.12.0",  # For Groq (ultra-fast inference)
  "aiohttp>=3.9.0",  # For Ollama (local models)
]
all = [
  "anthropic>=0.39.0",
  "google-generativeai>=0.8.0",
  "groq>=0.12.0",
  "aiohttp>=3.9.0",
]

[project.urls]
Homepage = "https://github.com/ShubhamSalokhe/nexuseval"
Issues = "https://github.com/ShubhamSalokhe/nexuseval/issues"


[tool.hatch.build.targets.wheel]
packages = ["src/nexuseval"]